+++ src/lib/services/expression/nftExpressionOld.ts
@@ -3,28 +3,30 @@ import { useMemo } from "react";
 import type { HexAddr, HexAddr32 } from "lib/types";
 import { isHexModuleAddress, isTxHash } from "lib/utils";
 
-export const useCollectionsExpression = (search = "") =>
+export const useCollectionsExpressionOld = (search = "") =>
   useMemo(() => {
     if (search.trim().length === 0) return {};
 
     return {
       _or: [
         { name: { _iregex: search } },
-        { id: { _eq: search.toLowerCase() } },
+        { vm_address: { vm_address: { _eq: search.toLowerCase() } } },
       ],
     };
   }, [search]);
 
-export const useCollectionActivitiesExpression = (
+export const useCollectionActivitiesExpressionOld = (
   collectionAddress: HexAddr32,
   search = ""
 ) => {
   const isNftAddress = isHexModuleAddress(search);
   const isHash = isTxHash(search);
 
-  const tokenIdSearch = isNftAddress
-    ? { nft_id: { _eq: search.toLowerCase() } }
-    : { nft: { token_id: { _iregex: search } } };
+  const tokenIdSearch = {
+    nft: isNftAddress
+      ? { vm_address: { vm_address: { _eq: search.toLowerCase() } } }
+      : { token_id: { _iregex: search } },
+  };
   const txHashSearch = {
     transaction: {
       hash: {
@@ -37,34 +39,15 @@ export const useCollectionActivitiesExpression = (
 
   return useMemo(
     () => ({
-      collection_id: { _eq: collectionAddress },
+      collection: { vm_address: { vm_address: { _eq: collectionAddress } } },
       ...(search ? { _and: searchOption } : {}),
     }),
     [collectionAddress, search, searchOption]
   );
 };
 
-export const useNftsExpression = (collectionAddress: HexAddr32, search = "") =>
-  useMemo(() => {
-    const orExpression = {
-      _or: [
-        { token_id: { _iregex: search } },
-        ...(isHexModuleAddress(search)
-          ? [{ id: { _eq: search.toLowerCase() } }]
-          : []),
-      ],
-    };
-
-    return {
-      collection: { _eq: collectionAddress },
-      is_burned: { _eq: false },
-      ...(search.trim().length > 0 ? orExpression : {}),
-    };
-  }, [collectionAddress, search]);
-
-export const useNftsByAccountExpression = (
-  accountAddress: HexAddr,
-  collectionAddress?: HexAddr32,
+export const useNftsExpressionOld = (
+  collectionAddress: HexAddr32,
   search = ""
 ) =>
   useMemo(() => {
@@ -72,17 +55,42 @@ export const useNftsByAccountExpression = (
       _or: [
         { token_id: { _iregex: search } },
         ...(isHexModuleAddress(search)
-          ? [{ id: { _eq: search.toLowerCase() } }]
+          ? [{ vm_address: { vm_address: { _eq: search.toLowerCase() } } }]
           : []),
       ],
     };
 
-    const collectionExpression = {
-      collection: { _eq: collectionAddress?.toLowerCase() },
+    return {
+      collectionByCollection: {
+        vm_address: { vm_address: { _eq: collectionAddress } },
+      },
+      is_burned: { _eq: false },
+      ...(search.trim().length > 0 ? orExpression : {}),
+    };
+  }, [collectionAddress, search]);
+
+export const useNftsByAccountExpressionOld = (
+  accountAddress: HexAddr,
+  collectionAddress?: HexAddr32,
+  search = ""
+) =>
+  useMemo(() => {
+    const orExpression = {
+      _or: [{ token_id: { _iregex: search } }],
     };
 
+    const collectionExpression = collectionAddress
+      ? [
+          {
+            vm_address: {
+              vm_address: { _eq: collectionAddress.toLowerCase() },
+            },
+          },
+        ]
+      : [];
+
     return {
-      owner: { _eq: accountAddress },
+      vmAddressByOwner: { vm_address: { _eq: accountAddress } },
       is_burned: { _eq: false },
       ...(collectionAddress ? collectionExpression : {}),
       ...(search.trim().length > 0 ? orExpression : {}),